New Fastest Route Option
You can check this option for isodistances/motor vehicle. Without this option the shortest path will be calculated, all route will be taken. Checking this option the fastest route will be taken and the polygon will be smaller, this is because usually the fastest roads are longer than the other roads